rearrange scaffolding for blog
