+++ app/work/articles/index.js
@@ -1,0 +1,9 @@
+import { buildingTemptations } from './building-temptations';
+import { samsungAmazonExperience } from './samsung-amazon-experience';
+import { gmbPraxisApp } from './gmb-praxis-app';
+
+export const articles = [
+  buildingTemptations,
+  samsungAmazonExperience,
+  gmbPraxisApp
+];
