+++ app/work/page.js
@@ -1,8 +1,11 @@
-import { categories, articles } from "./_assets/content";
+import { categories } from "./_assets/content";
+
 import CardArticle from "./_assets/components/CardArticle";
 import CardCategory from "./_assets/components/CardCategory";
 import config from "@/config";
 import { getSEOTags } from "@/libs/seo";
+
+import { articles } from "./articles";
 
 export const metadata = getSEOTags({
   title: `${config.appName} Blog | Stripe Chargeback Protection`,
